Adding a header comment
When I was reviewing the files, I was at a loss as to what this script did!  It only took me a minute or two to remember because I've used it recently but I imagine I will be more in the dark in a few months if I don't use it between then and now!  The comment will help me remember the use case for which I developed the script.