Adding `browser`
In the past, I loved using this little script I wrote on Windows called `noter`.
It would just copy stdin to a temporary file, launch `notepad` on it, and
clean up the file when finished.  I haven't had much use for it lately but I've
often wished for something like it for Linux that launched vi/vim.  So I finally
came up with one that works well!