Adding `pushsshkey`
I've used this on my Windows machine for a while but never added
it to git because I don't run it anywhere else.  I wanted to make
a change to it today to take care of the situation when the target
machine doesn't have a .ssh/ directory but I wanted to do it on
Linux using git so I added it.