Support `--logstash` option
I got sick of having to change the index name all the time when
I'm dealing with messages posted by logstash.  The format of
the index it creates is known so I taught this script the format
so I can just use an option to have it automatically generated
based on today's date.