Adding pound-bash to jwatch
When I tried to run the script on Windows, it was not finding
the location where my class was when I ran it as-is.  When I
specifically ran it with `bash` it worked fine.  I hope adding
a pound-bash to the script will work.  `/bin/bash` lives on all
of my Linux machines and cygwin on Windows.