Making table.py process stdin
I originally had made it so that if you called the script
directly, it would some some fake static data but yesterday
I had a case where I had some tabular data and I just wanted
to separate the columns with vertical bars to make it easy
to turn into a `makedown` table.  The data didn't even have
headings but I'm not sure that matters.  I just changed it to
process data from stdin if it isn't a tty.