Making `columns -s` the default
Too often when I use `columns`, I'm ending up using the `-s` option
to "squeeze" consecutive spaces.  I'm either always using it or the
use doesn't really matter because there aren't consecutive spaces in
the data.  I'm going to turn the option on as a default to squeeze
the spaces down to one.  I'll still leave the option around to turn
the behavior off if necessary.