Enhancing `beeper` a little
While using beeper today, I was a little confused when
it started because it wasn't clear to me that the command
I had issued before `beeper` had finished or not.  I could
hear the bepp but I was afraid of interrupting something
launched something in the background or something.  I didn't
think that was the case but I wasn't sure!

I decided to make `beeper` a little more explicit by printing
out a string before starting the beeping.  I also added
a handler for `ctrl-C`, printed a terminating string,
and added timestamps to both messages so you can see when
beeping started and when it ended.