Correcting a regular expression
When I tried running this on my Meerkat system, it failed
to parse the output properly.  It looked as though I expected
a regular expression to be anchored at the beginning of the
string being parsed but this was not the case on the Meerkat.
I'm not sure how this would have worked on other systems but I
think it's ok to remove the anchor. There's enough context to
match only what I want.