Fixing a unicode issue
I was seeing an exception when I tried to process JSON output from git commits that contained a backquote (`):

  ibmadmin@pfuntner1:~/bin$ toycurl commits | jsonflatten  >/dev/null
  Traceback (most recent call last):
    File "/home/ibmadmin/bin/jsonflatten", line 35, in <module>
      process(json.load(sys.stdin), [""])
    File "/home/ibmadmin/bin/jsonflatten", line 29, in process
      process(root[key], path + [str(key)])
    File "/home/ibmadmin/bin/jsonflatten", line 32, in process
      process(root[key], path + [key])
    File "/home/ibmadmin/bin/jsonflatten", line 32, in process
    process(root[key], path + [key])
    File "/home/ibmadmin/bin/jsonflatten", line 34, in process
    print "%s\t%s" % ('/'.join(path), stringify(root))
  UnicodeEncodeError: 'ascii' codec can't encode character u'\u2018' in position 714: ordinal not in range(128)
  ibmadmin@pfuntner1:~/bin$

I was at first using str(root) but I was trying other solutions and was trying to a new method I wrote that would be smarter about the type of object I was trying to turn into a string.  It turns out that using repr() was easier so I just switched from str() to that and got rid of my new method.