Allowing nonexistant store files to work
When I added the retry logic, I failed to account for
store files that have not yet been created.  I want
to be able to create a SecureKeyValues object with
a key even if the file doesn't exist.  Once you have
the object, you might add pairs and write the store
out to the file - that's how a store is created!

I had tried to add some logic to fail when the file
does not exist but that was a mistake - I was making
it impossible to ever create a store file.