Improving debugging to chase another bug
I think there's another bug because it's not working as expected
with the file I ultimately wanted to compare.  It understands that
the first few lines are common and that some lines after that are
unique to only one file but then it doesn't synchronize with the
rest of common lines in the file.  I still need to figure that out.