Fixing 'sync' logic
I had a logic problem in matching up common bits of a file
after the files didn't match.  If there the current lines
of each file did not not match but are found later on in
each companion file, I needed a better way to choose which
line should be considered "unique".  I wrote a new function
that considered how many matches could potentionally be made
in either direction.  The one who had more potential matches
would win and the line from the other file would be considered
"unique".