JSON output feature
I plan to reuse this script with another script and want the
other script to process output from this script in a JSON
object when invoked as a CLI.  Error messages are encapsulated
in a special attribute of the JSON output.  The script still
prompts for the key like usual EXCEPT when the store file
doesn't exist and the CLI 'read' operation is used.  In that case,
no key is prompted for and an error message indicating that the
fiel was not found is included in the JSON output.