Fixed a booboo
I was using `print` to print out the result I wanted to
turn that into a stream.write() so the newline was not
automatically added.  I committed a change that tried to
write to sys.stdin without even trying it out!

This uses the correct stream and I did try it first!