I should have been more persistent about trying out `isbinary`!
While adding the `--details` option to `isbinary`, I noticed
it was myteriously giving me bogus offsets for the binary data.
It turns out this was because I was examining the `details` of
the data buffer *after* I removed all the printable characters!
So of course the first offset is 0, etc.  This fixes that
problem.