Making columns unbuffered
I was trying to use `columns` with another command it seemed to
be taking a very long time.  At first, I thought one or both of
the commands were buffering output (both were Python) but that
wasn't the case.  The cause was that I had changed the first
command to sort the output based on a timestamp so it's buffering
the output internally and dumping it out at the end.  So it wasn't
really necessary to make `columns` unbuffered but I think it's a
good idea anyway.

For some reason, I had used "#! /usr/bin/env python" rather than
"#! /usr/bin/python" and I couldn't just add "-u" easily but I
don't think there's a reason to use `env` so I changed it to the
more common usage.