Moving list of tools to parent directory
As the list of tools being documented grew, I realized that the table I was
building with simple tool descriptions was losing the competition with the list
of files.  I merged the table from `doc/README.md` into the preant `README.md`
and am removing the table from the `doc/` directory.