Introducing new json tool
I was never completely satsifed with the json tool I had.  I think it was trying to do too much stuff.  So what I'm doing is:

  - rename the existing `json` script to `oldjson`
  - introducing a new `json` script to take its place.  It supports:
    -- traditional json.dump() style of output
    -- `flattened` style of output
    -- ability to dive to a particular element by path and ignore higher elements
    -- ability to `lop off` one or more levels below the resulting elements

I'm thinking of some sort of new filter that can do searches and return elements that either do or do not have a given value.  The new filter could be used first and then `json` could be used to display it.