Introducing jsonhunt and a *unit test*!
I've been messing with unit tests in my regular work and when I created this new script, I thought I would create unit tests for it.  Usually jsonhunt is used as a script from the command line but I turned it into a class so it could be instantiated from the unit test and driven from there.