Adding more changes for testing jsonhunt
In all honesty, the tests are failing right now but I don't have time to investigate.  I don't know if the tests are wrong (I think I may have dummied up an expected result completely until I knew what it would look like) or the jsonhunt script needs to be changed.  It could be a combination of both but I wanted to save my work so far so I can return to it anywhere.