Adding `-v` option to headgrep/tailgrep
I'm not even sure if this is going to be useful but I think I used a cool way so I figured I would save it.  I also thought about having some sort of extra option which I guess would depend on whether it's headgrep or tailgrep.  The usual behavior is:

  - headgrep: prints the top of the file through the first occurrence of the regexp
  - tailgrep: prints the last occurrence of the regexp through to end of the file

The new options would do:

  - headgrep --last: prints the top of the file through the *last* occurrence of the regexp
  - tailgrep --first: prints the *first* occurrence of the regexp through to end of the file

These new options are not part of this commit and I'm just thinking about them but this helps me think about new features.