Finalizing uniqc/extensions doc and fixing a bug
I moved to a different system and wanted to come up with a new example but was seeing some odd behavior from `uniqc`.  It turned out the the `squeeze` option was doing things I didn't want it to do.  I've turned the option off by default and will keep an eye on that.