Teaching htmlize about tab characters
I was using `htmlize` to try to _tablize_ a `*.tsv` file (I had downloaded a database query result set from Django) and tried to use:

  ```
htmlize -F \\t < file
  ```

But it didn't do what I wanted.  I figured the shell isn't passing in a tab character so I changed the script to use `'\t'` when `'\\t'` was specified as the separator character.