Adding --skip option to jsonjoin
Dealing with log output from Django tests, I was sick of having to modify the output to strip away non-log output.  The --skip option can be used to skip over data that doesn't appear to be part of a json dictionary or list.  I'm not sure it will work for general cases but it seemed to work great for the current problem so I want to commit the option into the script.