Changing numbering color
I was using a white background for some shell sessions and when I launched vi and turned on line numbering, the default coloring was a bright yellow which was nearly impossible to read.  I found a StackOverflow article about changing the color so I changed it to grey which shows up well on white or black.