Some changes to daterep
I was dissatisifed with daterep always replacing the timestamp with a fairly long string, sometimes longer than the original substring.  So I added some options to choose the format of the replacement string.  I also added an option to pad the replacement string if it was *shorter* than the original substring.

One bug is that I seem to have broken the ability to replace multiple hits on a line even when --single is not used.  I need to investigate that more.