Making fitwidth read lines in an unbuffered way
I was trying to combine fitwidth with another command like tailer today and realized it wasn't processing input as it was sent.  Looking more into it, it seemed that the way I was reading data was kind of naive and I didn't see any data until the the EOF was received.  I changed it to read and print out a line as soon as the newline is received.