Introducing --maxcols option for tabular
I discovered that the script didn't work very well on output from a command such as `env` when the value of one or more of the environment variables contained the column separator (an equal sign - I see them in LS_COLORS).  If you specify `--maxcols X`, the script will not produce more than X columns.  In the case of `env`, this combination of options would probably be appropriate:

	env | tabular --maxcols 2 --regexp =