Refining method to handle ctrl-C
My first version worked well on Windows but not quite as well on Linux where it looked like:

  $ pings localhost
  2018-11-30 14:51:51.807797 localhost UP   0:00:00.001989^C$

I would like a newline to be present so the prompt comes out nicely, especially in the case of the pings script which generates a lot of output but not a lot of newlines.  Then I was faced with the issue of Python lambda functions only taking a single *expression* (not a statement such as `print`).  I solved this by reusing the method in a sneaky way.  It's kind of recursive but will never call itself more than once.