In-progress changes to noter
These changes are working well for cygwin but not git bash since the /tmp directory is slightly different.  Additionally, the script is available on Linux and I was trying to add something so that it would fail gracefully on Linux since it's really designed for Windoze.  However, I could just have it use `vi -R` to view the file??.