Fixing `explode --count` for tarballs
The explode script works with either tarballs or zip files and I tested it with zip files but not with a tarball.  I had introduced a regular expression that the output had to match in order for a line of output to count as a file.  This is not necessary for tarballs so I changed to logic to only use the regular expression for zip files.