First phase of rework done
The script seems to accept `-?\d+(:-?\d+)?' and re=regexp well.  The approach I took makes one pass to select the lines and then a second pass to print the selected lines.  I would like an `immediate` style of operation too where rows could be duplicated if desired so I'll add that as an option.