Fixing localhost execution to use bash
When you did something like `chainssh ls -d /etc/\*`, it would just try to run it directly without resolving the wildcard because the shell was not being used.  I changed the code to use bash.

This is not an issue for a remote system because ssh is used and the bash (or whatever login shell) is used implicitly