Adding --empty option to json script
I had a case of using a yaml file that `json --flatten --yaml` didn't work well with.  It only listed a few elements when I knew there were many more.  The problem was that, for whatever reason, there weren't a lot of `leaf nodes` in the structure: there were many dictionaries that were empty but in the case of this yaml, the dictionaries were still significant.  So I introduced an `--empty` option that affects how the output is rendered when using `--flatten`.  I suppose it might be useful by other output methods too but I'll hold off introducing it to others for now.