Adding jsondiff and yaml tools
The yaml script is super simple but useful.  It isn't as robust as the json script but it produces JSON output so you could use the json script to filter its output if desired.

I also changed .gitignore to ignore a script I use at work but don't plan to check in