Handle more yaml files
I was using a regular expression to test for yaml files that was unable to determine that foo.yaml.orig was a yaml file.  I decided to relax the regular expression to make it so that the filename need not end in .yaml or .yml.  It must merely *contain* .yaml or .yml in the name.