Creating mongocli
This is a work in progress.  I started learning about mongodb hosted on the cloud and wanted to explore creating a cli without having to install a lot of stuff.  It does require the pymongo package for Python but I figure that will be necessary if I do anything in Python for mongodb on the cloud.