Replacing "#! /usr/bin/python" with "#! /usr/bin/env python"
I was trying to install a private version of Python on a Linux system to override the system version but was having trouble running some of my scripts because of the "hash bang" statement they were using.  I'm changing everything to use /usr/bin/env python to pick up whatever version is first in the PATH.

In at least one case, I think carriage returns were elimiated from the script - that's a good thing!

In some cases, the scripts were using "#! /usr/bin/python -u" to turn off stream buffering but I think I set the PYTHONUNBUFFERED environment variable on most of my users/systems so the option is not necessary anymore.