Catching error when encryption package is not installed
I want to be able to use SecuryKeyValues.py from a system that doesn't have the cryptography package installed but I don't want it blowing up completely.  It can issue messages (which are only shown when --verbose is used) but if the package isn't installed, it will behave as though no keys were set.