Improving ssh key support
I had used just a single key (userid) in the secure store but realized that I should really use a multi-key structure of server/cluster/user.  SecureKeyValues.py already supported this (for the most part!) so I changed mongocli to ask for the password using that hierarchy.

Most of the changes were in SecureKeyValues.py:
  - I had a bug in setting keys in a hierarchy such as I was doing so I fixed that
  - I changed the CLI interface so the --key and --ssh options were mutually exclusive
  - I had some bugs in logging messages so I took care of those