Improving imcomingHost
While running `incomingHost`, I noticed an exception when `nslookup` was not available on a system.  I switched the script to use argparse/logging and added some logic to catch the exception.

I also noticed I had added a debugging message to SecureKeyValues.py which I guess I neglected to check in so I'm doing that now.