Fixing mistake in arguments for options
I had mistakenly used `nargs='*'` for options such as `-c CLUSTER`.  I suppose it would have accepted multiple arguments but I never used it that way and did not intend for it to be used that way.  The option is optional but the argument is not if the option is expressed - one and only one argument is expected.