Fixing an error message
I was using `sys.stderr.write()` to write an error message directly to the screen but had neglected to place a newline to the end of the message.  Rather than add a newline and since I had already started using the logging module, I decided to use `log.critical()` to display the error message.