Creating nearcompare
This is an attempt to come up with a script that is able to take to files and show how similar they are on a gross level.  For instance, if you give it two text files, it will read the files and display how many lines, words, and bytes each file has.  It makes no determination about how similar the files are but the user can use the information to judge for themselves and take further actions if desired.