WIP: Local side of tarcopy
I had an idea of a copy command that would work with using `tar` on both ends - something that could work even if the destination is a remote system via ssh.  The local side seems to be working well but I haven't done anything for a remote system really yet.