Redoing `needsdoc` again
At one time, I think `needsdoc` was just relying on the filesystem.  Then I changed it to rely on the files as commited (via `git log`).  Today I realized a flaw with this is that `git log` has names of files that have been removed or renamed to something else.  I switched it back to rely on the filesystem and also sorted the list by modification time of the files.