Adding doc for fulltime
I also realized I had a copy of the paramiko script in the root of the repo.  At first I thought I had misplaced it and it should have been in bin/ but there was already a more recent verison in bin/.  So I'm not sure how that happened but I'm getting rid of the old one from the root.