Fixed reading supported!
This was easier than I thought!  The approach I used was to find the beginning and ending columns of all lines based on whether all the rows had a blank in the column in question.  I'm not sure it will work in all instances but it's working for the df command right now so I'm quite happy!  For example:

  $ df . | supertable -i fixed -o json --headings
  [
    {
      "1K-blocks": "1922727728",
      "Available": "1833812760",
      "Filesystem": "/dev/sda1",
      "Mounted on": "/media/mrbruno/ExtraDrive1",
      "Use%": "5%",
      "Used": "88898584"
    }
  ]
  $