Protecting embedded blanks
On Windoze, I sometimes have a user with an embedded blank such as "first-name last-name" (/home/John Doe/.profile) which can cause problems if you're not careful and just try to do:

  vi $(bashprofiles)

You might have to use double quotes around the command substitution (*not* single quotes!).  This is an attempt to work around that issue by automatically printing quotes around the path when there are embedded blanks.  This might not work in some cases (/home/John O'Doe/.profile) but will work for me, I think.