Adding note on bashprofiles
I tried to fix an issue dealing with an embedded blank in the path name for a profile but I haven't found a good general solution for the script yet.  For now, the best solution is to be aware of the issue and use quotes to protect the blanks.  I added a note on this an example where protection is needed and proof that protection works.