Creating decolor script
I was using the help from the aws cli and piping it into more/br - Even though it wasn't writing to the terminal, it used data sequences (eg '\x1b[0m') to make use of various colors, making it awful to read.  This is a simple script that just removes all of those sequences.