PS1 improvements for current branch
I was dissatisified with the "current branch portion" that I recently introduced to my PS1.  I was on Unix and did a `sudo bash` and got an error producting the current branch.  My original approach was to use a bash function but since the bash function wasn't present in the root process, it was not found.  Here's what I did:

  - The currbranch script will be used
    - The script won't produce an error message if a branch is not found and -q or --quiet is specified to the script
    - I adjusted the output so that when the current branch is found, a blank proceeds the branch if stdout is not the terminal
  - misc/.bashrc:
    - I use the absolute path of the currbranch script
    - If the script is not found for some reason, I suppress any errors by using the `true` utility
  - Initially, I was using lspath to help me figure out why the script wasn't found.  It was producing an undesirable error because I was looking for a function.  I changed the script to handle this error more gracefully.