Restroing Popen.wait()
When I recently changed this script to work on multiple systems as once, I removed an exit() which was also doing the Popen.wait().  When I removed the exit(), nothing was calling Popen.wait() anymore.  This needs to be done so the script doesn't exit or move on to the next system too quickly.