The script I used to change `python` to `python2` in my files.
I believe this is the first Python 3 script I'm committing!  This was the one where I wanted to make use of `map()` to drive my `dive()` method but rather than figure out how to do it with Python 3's weird concept of `map()`, I decided to just bust out a little loop.