Improving the regular expression checking a little
I tried doing `dowhile -i '' ...` because I didn't really care what the output or exit status was but it didn't pass the check to make sure either -i/-x or -s/-f were specified.  Maybe I need a `forever` option for such a situation but I just changed the script to allow to accept '' as a regular expression (which it wasn't doing).