Improvements on ansible-stat
I improved the IP address column so it's a real external address.  This goes for both remote hosts and localhost!

I also added an option to makeawshosts to force the remote user name if it can't be discerned by the remote system name.  I think the AWS API will return the OS information so that should be sufficient as well.