Getting instance names properly
It became necessary for me to add tags to my instances which exposed a false assumption about the JSON I was getting back from aws.  The instance name is yet another tag and is not not necessarily the first one.  So when there are other tags, you have to hunt down the `Name` tag and use its value.