Creating new tool
I am displacing the old repeat script (I'm not sure what the use was, exactly), calling it repeat.py, and creating a new repeat script that can repeat commands.  There are native commands that could do the same thing as the script such as:

  foo | xargs -I {} bar {}

but I wanted a script.  It also has logging and a --dry-run flag.