Creating new aws/gcp script
Right now, it just displays and shows the image it was based on.  It might be a little slow and I could work on that.  It does a separate call for each AWS image (as long the image isn't in an internal cache). This could be improved if there was a single call doing all the images but the script would probably have to backfill the image name.