Format assertion text properly
Really, this entire script could use a facelift with the logging and argparse modules but I'm not going to do that right now.  The only reason I even triggered the assertion failure is because my home directory was the parent of a git repo - it was not **in** a git repo.