Improving Windoze ssh frontend
I set up a Ubuntu 18 virtual instance and found the frontend was not setting the Windoze title to the destination automatically like on other distros.  I wanted to use my `windoze_title` tool to set the title before invoking `ssh` (it already resets the title after `ssh` returns).  I had to set up the script to accept all the Windoze options and reform them so I could distinguish the destination from options and the command.