Creating tool to ping a host using Ansible
The basic ping command is great but I've had experience with virtual hosts that can't be pinged with it.  This script uses the Ansible ping module.

I provided a few options to control the behavior but didn't go out of my way too far to align it with the ping option and see if there's a 1-to-1 correlation.