Enabling --become option
I can't believe I never wired in this option!  It's parsed but I didn't do anything with it!

Also, I noticed that the brain-dead Python3 temporary file class was forcing me to write bytes to the file until I overrode the mode so it wasn't in the insane binary mode.