Adding --strict_headings option
For a while, I've been hoping for a alternate style of parsing columns when --fixed and --headings are used together.  Originally, the default is to look for the first column of all rows where there's a blank - that's the end of the column.  That causes problems for an example like this:

  head 1    head 2     head 3
  cell 1 1  cell 1 2   cell 1 3
  cell 2 1  cell 2 2   cell 2 3

The parsing is off and if the output is JSON, it comes out as:

  [
    {
      "": "3",
      "1": "1",
      "2": "1",
      "3": "1",
      "head": "cell"
    },
    {
      "": "3",
      "1": "2",
      "2": "2",
      "3": "2",
      "head": "cell"
    }
  ]

I added a --strict-headings option that only pays attention to the row with the headings - it doesn't look at the data to determine where the columns begin and end.  This does require *TWO* *OR* *MORE* blanks to separate columns in the heading but this can be better than --headings:

  [
    {
      "head 1": "cell 1",
      "head 2": "1  cell 1",
      "head 3": "2   cell 1 3"
    },
    {
      "head 1": "cell 2",
      "head 2": "1  cell 2",
      "head 3": "2   cell 2 3"
    }
  ]

This can be a much better alternative.