Improving gitlog for remote repos
I wanted to add an option for remote repos but realized I already had one with kind of an obscure name.  Regardless, it didn't work as intended because it only looked at the master branch of the remote repo so I fixed that and gave the option a less-obscure name.

This was an old script and I probably would do it a little differently if I did it from scratch today but it's fine.