Adding synonym of -u for --user
I think when I was first writing this I was going to use another option instead of --user which would have made it impossible to define the short name as well.  I don't remember what it was but this is often the case when think about defining a -h option because it would conflict with --help.  I don't think it was --help but that's the kind of conflict I think I was trying to avoid.  Since -u won't conflict with anything else, I'll use it.