Fixing bug when renaming filenames
I realized that when whitespace is inside a filename that gets renamed (yes, I did that once - the rename was to GET RID of the white space), git is not as nice about for
matting the file names as possible.  But it does use tabs between the tokens so that's good:

  R100^Iimages/red x.png^Iimages/red_x.png$

I can parse out things using the tabs.