Changing args.files/sys.stdin behavior a little
I didn't like that the tool would not work with xargs:

  find ... | xargs br

This would complain because stdin was redirected and there were files on the command line.  I made it so that if there are files on the command line, we use those regardless of stdin.

If there are no files on the command line, the tool requires that stdin be redirected.