Creating tool to clean up ~/.ssh/known_hosts
I create many AWS/GCP cloud instances that aren't very long lived.  I usually add their SSH fingerprint to ~/.ssh/known_hosts and try to remove the fingerprint when I remove the instance but am not always so good about that.  This script looks at all the lines in the file, determines which belong to AWS or GCP, and remove them from the file.

I designed the script to remove all the fingerprints, even for instances that are actively in use because `add-to-knownhosts` can add them back in very easily.