Creating new tool
This uses a lot of my other tools under the covers but works pretty well, I think.  I hard originally done it in bash but when I wanted to add the number of tasks that have run, I decided to switch to Python and commit it to my toys repo.