A little housekeeping
The script was still importing the getopt module which made me believe it made use of it but it didn't.  I had converted it to use argparse but didn't bother getting rid of the import.  I also moved the filename column to the front since that was kind of the big feature of this tool.