Redesign
I learned about using the `docker run --detached ...` option, `docker attach ...` to get to the container, and using `CTRL-p CTRL-q` to detach again.

I think this could present problems for a script I have that uses docker-start but I like this method better.  It's cleaner, doesn't rely on a sleeping process, and allows for you to disconnect and later connect to the container.  You can also do `docker exec ...` to start a brand new shell in the container although it's not detached.