A couple of tweaks
I wanted to include the overall data position and also didn't see a need for the option to report on the line number:

  - I added logic to keep track of the overall data position
  - I removed the --line option
  - I removed the ability to print stats WITHOUT the line number
  - I added the overall data position to the stats

I'm also working on a new arbitrary JSON tool and renamed the existing `jsons` script to `jsons.py`.  But I haven't commited the new `jsons` script yet.