Creating new tool
I had it in the back of my mind to parse the output of 'diff --side' but could never nail it down.  I have a simple approach that I think will work so I want to try it out.  It worked for:

  [ubuntu@pfuntner-ubuntu CCC-1527]$ sidediff <(echo -e 'a\nb\nc') <(echo -e 'a\nc')
  a                                                               a
  b                                                             <
  c                                                               c
  [ubuntu@pfuntner-ubuntu CCC-1527]$ sidediff <(echo -e 'a\nb\nc') <(echo -e 'a\nc') | side-diff-parse
  a||a
  b|<|
  c||c
  [ubuntu@pfuntner-ubuntu CCC-1527]$ sidediff <(echo -e 'a\nb\nc') <(echo -e 'a\nc') | side-diff-parse | table -i sep --reg \\\| --sep \ \|\  -o fixed
  a |   | a
  b | < |
  c |   | c
  [ubuntu@pfuntner-ubuntu CCC-1527]$

It automatically seeks out a seperator that isn't used in the input.  The first one it tries is the vertical bar which worked so it used that.