Defaulting to the current working directory
I am really liking the --all option and was trying to use it for the current working directory which is part of a git structure.  I had made it so that you had to specify '.' or something else regardless but I think it makes sense to default to the current directory.