Reworking driving script, allowing localhost
I had a need to use the script for a Docker container.  I thought I could make use of a Docker connection but the driving script was just bash so it was a little too stupid to do what I wanted so I converted it to Python 3.

The Docker connection wasn't sufficnet because I wanted to put this inside a Dockerfile and I didn't use the Docker connection after all for what I was doing.  Regardless, I think the ability is still good to have.

I ended up copying the files to the temporary container and running the playbook for `localhost`.

The final problem was that my playbook broke when the ssh keys weren't available.  I solved it by introducing a variable to skip copying.  Perhaps it could be done differently but it's working this way for now so I'm happy.