Making use of args.targets again
I think when I refactored this script to make use of the gityup class, I stopped using targets specified on the command line.  These are used to only find commits that affect the specified targets.

I restored the use of args.targets and the class focused on commits without any fuss!