Adding --channel option
I ran into a situation where there were two possible channels but one was "false" and I really wanted to use the "true" one.  I suppose the script might be smart enough to figure out which is more "true" but I just added an option to explicitly select the channel expressed as a column number.  This might require execution of the tool at least once with --verbose to see the potential channels but it worked well for me in this situation.