Tweak
I ran into a situation when I wanted the script to use the shell the execute the target.  I'm not sure why this was necessary now but I think I might have been testing some kind of shell script (as opposed to a binary executable) that doesn't have a hash-bang (#!).  I can't remember what it was I was trying what I was using as a target but that's what I think was going on.