Correcting log object
I was trying to make use of instances.py as a class in another script and discovered there are some errant statements trying to use the log object rather than self.log.  When instances.py is used from the command line, the log object is available.  But when it's used as a class, there is no log object and self.log should be used instead.