Tweak to backfill gcp disk name
When I was making use of the Cloud9 terraform role to create an instance, I had trouble detecting the version of the OS.  I made some tweaks so it would work after the terraform role.  I also created an instance using my old method and it worked so I'm committing the changes.