Improving --flatten sorting
I was dissatisfied with the order in which the paths appear when you use --flatten.  I thought I had tried to sort list indices numerically and there was code for it but a call to the sorting method was commented out for some reason.  I kind of remember doing that but when I uncommented it and tried a structure, it looked great and there were no errors.

I'm commiting the change with the call to the sorter method uncommented.