Adding more potential delimiters
I was having trouble processing a side-diff between two files because all of the potential delimiters were used in the data.

There are a couple of more normal punctuation characters that I wasn't using so I added them but that didn't help.

I used the Windoze character map app to enter a "¦" (a broken vertical bar) and that was not used in the files so I'm changing the script to use that as a potential delimiter too.