Changing how work laptop is detected
I had an `if` statement that used the hardcoded system name of my work laptop which was probably a poor choice.  When I got a new laptop, the system name was different.

I didn't experience a problem but I decided to change the `if` statement to decline to run for `instances.py --make` anytime it's run from a Windoze system.