Adding debugging message
Whe installing to a cloud machine via ssh, the command was being `lost` in the debugging messages because it was base64-encoded in order to preserve metacharacters.  I added a message to reveal the command in the clear:

  2022-10-15 07:41:31,261 DEBUG /home/mrbruno/bin/install-python:55 Encoding 'python -c "exit(42)"' for the trip!
  2022-10-15 07:41:31,261 DEBUG /home/mrbruno/bin/install-python:59 Executing ['ssh', 'localhost', '--', 'echo cHl0aG9uIC1jICJleGl0KDQyKSI= | base64 -d | sudo -i']

The first message shows the original command `python -c "exit(42)"`.  `cHl0aG9uIC1jICJleGl0KDQyKSI=` is the base64-encoding that protects embedded commands and parens across ssh & sudo.