Creating doc for megassh command
I started this at home where I don't have access to a system to which I can ssh.

  - I'm dissatisfied with the bevy of Cloud promises for "free" systems.  Sure, they're free for a while but their trial periods all expire after some period.  I would like a cloud provider to come along and offer free instances with some conditions.  I don't need the instances for long or for heavy use!
  - I tried bringing up a VirtualBox and I think I got close to being able to ssh to it but didn't seem to like my ssh key and it kept telling me `Permission denied (publickey)`

I tried to do all the other doc except for the examples.  For now, I'm going to stash the doc into a separate branch until I get a chance to complete the examples.