Hacks to allow stdin
I haven't used this script in a while but when I tried to use it today by piping through stdin, it failed because stdin is in text mode, not binary.  I figure I made some changes a while back that allowed for binary files expressed in paths on the command line but I ignored stdin.