Major long-overdue rework!!
I've neglected changes to this script for too long.  It's useful in general but I think I stopped using it because it had problems.  I dusted it off, improved a lot of things, and made sure it was working.  Along some of the work:

  - Switched from Python 2 to Python 3
  - Used byte strings for encryption and writing to the store file
  - Switched from using locals() in messages to f-strings
  - Improved logging
  - Added a list option
  - Forced stdout redirection for get operations - I would like to think I'll have the presence of mind to pipe it into `br` when I use it
  - Switched from double-quoted strings to single-quoted - that's my style these days
  - Enforced mode 0600 for the store file

I verified that the `test` operation worked well.  I can also use the `list` operation on the test store as long as I use the correct key.